+++ drivers/media/platform/qcom/iris/Makefile
@@ -4,13 +4,16 @@ qcom-iris-objs += iris_buffer.o \
              iris_ctrls.o \
              iris_firmware.o \
              iris_hfi_common.o \
+             iris_hfi_gen1.o \
              iris_hfi_gen1_command.o \
              iris_hfi_gen1_response.o \
+             iris_hfi_gen2.o \
              iris_hfi_gen2_command.o \
              iris_hfi_gen2_packet.o \
              iris_hfi_gen2_response.o \
              iris_hfi_queue.o \
-             iris_platform_gen2.o \
+             iris_platform_vpu2.o \
+             iris_platform_vpu3x.o \
              iris_power.o \
              iris_probe.o \
              iris_resources.o \
@@ -25,9 +28,5 @@ qcom-iris-objs += iris_buffer.o \
              iris_vpu4x.o \
              iris_vpu_buffer.o \
              iris_vpu_common.o \
-
-ifeq ($(CONFIG_VIDEO_QCOM_VENUS),)
-qcom-iris-objs += iris_platform_gen1.o
-endif
 
 obj-$(CONFIG_VIDEO_QCOM_IRIS) += qcom-iris.o
